+++ shelf_packages/open_api_specification/lib/src/specs/schema.dart
@@ -115,11 +115,15 @@ class SchemaOpenApi extends OriginalJson implements SchemaOrRef {
   @JsonKey(readValue: _readAdditionalProperties)
   final SchemaOrRef? additionalProperties;
 
+  @JsonKey(toJson: $nullIfEmpty)
   final List<SchemaOrRef>? allOf;
 
   // final List<SchemaOrRefOpenApi> anyOf;
 
-  // final List<SchemaOrRefOpenApi> oneOf;
+  @JsonKey(toJson: $nullIfEmpty)
+  final List<SchemaOrRef>? oneOf;
+
+  final DiscriminatorOpenApi? discriminator;
 
   /// With [TypeOpenApi.object]
   final List<String>? required;
@@ -165,6 +169,8 @@ class SchemaOpenApi extends OriginalJson implements SchemaOrRef {
     this.additionalProperties,
     this.propertyNames,
     this.allOf,
+    this.oneOf,
+    this.discriminator,
     this.required,
     this.nullable = false,
     this.default$,
@@ -189,6 +195,19 @@ class SchemaOpenApi extends OriginalJson implements SchemaOrRef {
       _$SchemaOpenApiFromJson(OriginalJson.wrap(map));
   @override
   Map<String, dynamic> toJson() => _$SchemaOpenApiToJson(this);
+}
+
+@SpecsSerializable()
+class DiscriminatorOpenApi {
+  final String propertyName;
+  final Map<String, String> mapping;
+  // final String defaultMapping;
+
+  const DiscriminatorOpenApi({required this.propertyName, required this.mapping});
+
+  factory DiscriminatorOpenApi.fromJson(Map<String, dynamic> map) =>
+      _$DiscriminatorOpenApiFromJson(map);
+  Map<String, dynamic> toJson() => _$DiscriminatorOpenApiToJson(this);
 }
 
 class SchemaRef extends SchemaOrRef {