+++ shelf_packages/shelf_open_api/lib/shelf_open_api.dart
@@ -37,3 +37,9 @@ class OpenApiRouteMount {
 
   const OpenApiRouteMount(this.serviceType);
 }
+
+class OpenApiDiscriminator {
+  final Object? value;
+
+  const OpenApiDiscriminator([this.value]);
+}